refactor: progress 폴더명 변경
