+++ src/components/detail/DetailParticipants.tsx
@@ -3,8 +3,7 @@ import React from "react";
 import { cn } from "@/utils/detail/cn";
 import { ParticipantsList } from "./ParticipantsList";
 import { TParticipatedUserClient } from "@/types/supabase/supabase-custom.type";
-
-import { ParticipantsProgressA } from "./ProgressAnimation/ParticipantsProgress"
+import { ParticipantsProgressA } from "./progress/ParticipantsProgress"
 
 interface IDetailParticipantsProps {
   participants: TParticipatedUserClient[];
@@ -25,7 +24,7 @@ export const DetailParticipants: React.FC<IDetailParticipantsProps> = ({
       className={cn("relative flex flex-col gap-4 px-5 py-5 mt-5 lg:px-6 lg:py-8 lg:mt-8 bg-background400 rounded-2xl items-center", className)}
     >
         {/* 참여 인원수 Progress Bar */}
-        <ParticipantsProgressA 
+        <ParticipantsProgressA
           currentCount={currentParticipants}
           minParticipants={minParticipants}
           maxParticipants={maxParticipants}
