+++ src/components/detail/progress/DetailProgress.tsx
