+++ src/components/detail/progress/ParticipantsProgress.tsx
