+++ src/components/detail/progress/Progress.tsx
